refactor: move preamble & output cfg to own files
